requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar


Appium_Python_Client == 0.25

python_dotenv == 0.8.2

scikit_image == 0.13.1

selenium == 3.6.0

pytest == 3.5.0